Remove glfw input and replace with sdl input events
Window input now comes from sdl events.

+++ AlphaEngine/Include/HID/HIDTypes.h
@@ -27,9 +27,9 @@ namespace alpha
     */
     struct HIDAction
     {
-        HIDAction(unsigned short actionRaw, std::string actionName);
+        HIDAction(unsigned int actionRaw, std::string actionName);
 
-        unsigned short raw;
+        unsigned int raw;
         std::string name;
     };
